Refactor to trait
