+++ src/Builders/Statements/EloquentStatementBuilder.php
@@ -3,12 +3,15 @@
 namespace Fidum\BlueprintPestAddon\Builders\Statements;
 
 use Blueprint\Models\Statements\EloquentStatement;
+use Fidum\BlueprintPestAddon\Builders\Concerns\ModelStatementHelper;
 use Fidum\BlueprintPestAddon\Builders\PendingOutput;
 use Fidum\BlueprintPestAddon\Enums\Coverage;
 use Illuminate\Support\Str;
 
-class EloquentStatementBuilder extends ModelStatementBuilder
+class EloquentStatementBuilder extends StatementBuilder
 {
+    use ModelStatementHelper;
+
     /** @var EloquentStatement */
     protected $statement;
