+++ src/Builders/Statements/ValidateStatementBuilder.php
@@ -6,14 +6,16 @@ use Blueprint\Models\Column;
 use Blueprint\Models\Controller;
 use Blueprint\Models\Model;
 use Blueprint\Models\Statements\ValidateStatement;
+use Fidum\BlueprintPestAddon\Builders\Concerns\ModelStatementHelper;
 use Fidum\BlueprintPestAddon\Builders\PendingOutput;
 use Fidum\BlueprintPestAddon\Contracts\TestCaseBuilder;
 use Fidum\BlueprintPestAddon\Traits\PopulatesTestStub;
 use Illuminate\Support\Str;
 use Shift\Faker\Registry as FakerRegistry;
 
-class ValidateStatementBuilder extends ModelStatementBuilder implements TestCaseBuilder
+class ValidateStatementBuilder extends StatementBuilder implements TestCaseBuilder
 {
+    use ModelStatementHelper;
     use PopulatesTestStub;
 
     /** @var ValidateStatement */
